# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Claimbase Expense Claim Management System API
servers:
  - url: /api
    description: Base API path
security:
  - TelegramAuth: []
...
paths:
...
  /users/{userId}/claim-periods:
    get:
      operationId: listClaimPeriods
      tags: [claims]
      summary: List claim periods for a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
        - name: companyId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of claim periods
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClaimPeriodResponse"
    post:
      operationId: createClaimPeriod
      tags: [claims]
      summary: Manually create a claim period
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClaimPeriodBody"
      responses:
        "201":
          description: Created claim period
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClaimPeriodResponse"
  /claim-periods/{periodId}/generate:
    post:
      operationId: generateClaimForm
      tags: [claims]
      summary: Generate claim form (PDF/Excel)
      parameters:
        - name: periodId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateClaimFormBody"
      responses:
        "200":
          description: Generated claim form URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateClaimFormResponse"
  /claim-periods/{periodId}/download/{format}:
    get:
      operationId: downloadClaimForm
      tags: [claims]
      summary: Download claim form
      parameters:
        - name: periodId
          in: path
          required: true
          schema:
            type: integer
        - name: format
          in: path
          required: true
          schema:
            type: string
            enum: [pdf, excel]
      responses:
        "200":
          description: File download
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
  /claim-periods/{periodId}/submit:
    post:
      operationId: submitClaimPeriod
      tags: [claims]
      summary: Toggle claim submission status
      parameters:
        - name: periodId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated submission status
          content:
            application/json:
              schema:
                type: object
                properties:
                  periodId:
                    type: integer
                  submittedAt:
                    type: string
                    nullable: true
                  message:
                    type: string
...
components:
  schemas:
...
    ClaimPeriodResponse:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        companyId:
          type: integer
        periodLabel:
          type: string
        periodStart:
          type: string
          format: date-time
        periodEnd:
          type: string
          format: date-time
        status:
          type: string
        pdfUrl:
          type: string
          nullable: true
        excelUrl:
          type: string
          nullable: true
        baseCurrency:
          type: string
        totalAmount:
          type: string
          nullable: true
        downloadedAt:
          type: string
          format: date-time
          nullable: true
        submittedAt:
          type: string
          format: date-time
          nullable: true
        receiptCount:
          type: integer
          nullable: true
        companyName:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - companyId
        - periodLabel
        - status
        - baseCurrency
    CreateClaimPeriodBody:
      type: object
      properties:
        companyId:
          type: integer
        periodLabel:
          type: string
          description: "Period label in YYYY-MM format"
        periodStart:
          type: string
          format: date
        periodEnd:
          type: string
          format: date
        receiptIds:
          type: array
          description: "Optional list of specific receipt IDs to include (for partial claims). If omitted, all unclaimed receipts for the period are included."
          items:
            type: integer
        rateOverrides:
          type: array
          description: "Optional rate overrides for receipts before claim generation"
          items:
            type: object
            properties:
              receiptId:
                type: integer
              rate:
                type: string
            required:
              - receiptId
              - rate
      required:
        - companyId
        - periodLabel
        - periodStart
        - periodEnd
    GenerateClaimFormBody:
      type: object
      properties:
        format:
          type: string
          enum: [pdf, excel, both]
        rateOverrides:
          type: array
          items:
            type: object
            properties:
              receiptId:
                type: integer
              rate:
                type: string
            required:
              - receiptId
              - rate
      required:
        - format
    GenerateClaimFormResponse:
      type: object
      properties:
        periodId:
          type: integer
        pdfUrl:
          type: string
          nullable: true
        excelUrl:
          type: string
          nullable: true
        status:
          type: string
      required:
        - periodId
        - status
...
  securitySchemes:
    TelegramAuth:
      type: apiKey
      in: header
      name: X-Telegram-Init-Data
      description: Telegram Web App initData string (HMAC-verified, 1-hour TTL)
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT session token from Telegram Login Widget auth
    InternalAuth:
      type: apiKey
      in: header
      name: X-Internal-Key
      description: Server-to-server authentication key (used by Telegram bot only)